% Auto-generated: do not edit by hand
\name{dashJSMessageReceiver}

\alias{dashJSMessageReceiver}

\title{DashJSMessageReceiver component}

\description{

}

\usage{
dashJSMessageReceiver(id=NULL, conditions=NULL, data=NULL)
}

\arguments{
\item{id}{Character. }

\item{conditions}{Named list. }

\item{data}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

