% Auto-generated: do not edit by hand
\name{dashJSMessageSender}

\alias{dashJSMessageSender}

\title{DashJSMessageSender component}

\description{

}

\usage{
dashJSMessageSender(id=NULL, data=NULL, origin=NULL, target=NULL)
}

\arguments{
\item{id}{Character. }

\item{data}{Named list. }

\item{origin}{Character. }

\item{target}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

